[REFACTOR] Certik Audit Recommendation - Exhibit 23
Description:
The functions “getCurrentPayoutPeriod” and “_payoutPeriodFromCycleAndPeriodLength” are
identical with the sole difference between them being the arguments they consume.

Recommendations:
We advise that “getCurrentPayoutPeriod” internally calls
“_payoutPeriodFromCycleAndPeriodLength” with the correct arguments to avoid unnecessary
duplication of code and reduce the total bytecode of the contract. Additionally, it appears that
the second argument of “_payoutPeriodFromCycleAndPeriodLength” is always the
“_payoutPeriodLength” member of the contract and as such could be avoided altogether,
meaning that the function “_payoutPeriodFromCycle” between lines 280 and 282 can also be
removed.