[REFACTOR] Certik Audit Recommendation - Exhibit 25
Description:
The code block between lines 209 and 245 contains multiple “if” statements that could be
simplified by following the previous Exhibits regarding memory management as well as the
optimization of the “if” clauses contained in this exhibit.

Recommendations:
The “if” conditional on line 211 evaluates the expression “snapshot.cycleRangeEnd !=
currentCycle - 1”. If it is true, the body of the “if” clause is executed. In that body, the assignment
“snapshot.cycleRangeEnd = currentCycle - 1” immediately follows on line 212. This means that
if “snapshot.cycleRangeEnd != currentCycle - 1” evaluated to “true”, it will now evaluate to
“false”.

On line 222, the “cycleRangeEnd” member of the “snapshot” struct is changed to an arbitrary
value. However, on the “if” clause in line 225 the new value is checked against the previous one
and if they are different, the “if” clause’s body executes. This means that within the body of the
“if” clause located in line 225 the “snapshot.cycleRangeEnd != currentCycle - 1” expression will
evaluate to “true”.

As “snapshot.cycleRangeEnd != currentCycle - 1” is equivalent to “snapshot.cycleRangeEnd + 1
!= currentCycle”, the “if” clause of lines 230 - 232 can be moved within the body of the “if” clause
of line 225 and the “snapshot.cycleRangeEnd + 1 != currentCycle” check can be skipped.

In general, the code block of the “_getOrCreateLatestCycleSnapshot” should be refactored as it
is highly convoluted at its current state and unnecessarily so.