[CHANGED] Certik Audit - Exhibit 12 (round 3)
[INFORMATIONAL] Description:
The loop iterates either until the maximum periods to calculate or the current period is met by “claim.periods” and “nextClaim.period” respectively.

Recommendations:
As both variables are incremented by 1 on each iteration, it is possible to simplify the “while” condition to a single logical calculation as well as a single variable increment. The minimum of “currentPeriod” minus “nextClaim.period” and “maxPeriods” is the number of iterations the loop will go through and what “claim.periods” will hold.

As such, one could assign this value to “claim.periods” directly and then instead conduct the “while” loop as long as “nextClaim.period” is not equal to either “currentPeriod” or “nextClaim.period” plus “maxPeriods”, depending on the result of the previous computation. This would reduce the number of computations from two to one and the number of logical comparisons from two to one.